requirements.txt
pynest-api==0.1.2
aiosqlite==0.19.0
pbr==5.11.1
